fix(mesh): a factor is judged at the width the document wrote it in
Four reviews against this branch, and the checks were green throughout.

The reader clamped where its own documentation promises it refuses. A
factor's range was checked after narrowing the document's number to the
width this crate keeps, and a value one unit in the last place past the
maximum narrows to exactly the maximum - so the check could not see it,
the document was not conforming, and the caller got the limit with
nothing said. The whole argument of this module is that the format's
ranges are stated rather than conventional, so they are refused rather
than clamped; the ordering inverted that in the one place the difference
is invisible. Checked wide, then narrowed.

A material, and its shading half, are asked their kind rather than
merely their presence. Every member lookup answers absent on a number, a
string or an array, so a table of those read as a table of materials
that each said nothing, and this reader's own rule then handed back a
full default material for a document that described none. An
out-of-range factor one array wrapper deep read clean the same way.

The material table has the ceiling this crate's amplification rule asks
for. Geometry is capped because a megabyte of legitimate input built
fifty-eight megabytes of it; a two-byte array element is a whole
material, which amplifies harder, and the table had no equivalent.

Pairing a mesh with its surfaces is one pass. Asking per primitive walked
both tables from the head each time, which is the cost this file
converted three other readers away from, and the singular form's own
documentation invited exactly that loop.

A texture index and a material index are bounded by the table they name.
Neither is resolved - this crate does not read the texture table - and
the documentation says that rather than implying more. Passing them
through unchecked was the one thing this reader's own module doc calls
the commonest fault in a hand-edited document.

An alpha mode the format does not define has its own refusal, carrying
what was spelled. It was being reported as a construct the format defines
and this reader does not, which told the caller to convert a file that
needed repairing.

A colour array is exactly as long as the format states. Reading the first
few and dropping the rest hid whatever the tail said, including
components that would have been refused. A cutoff is range-checked
whatever the mode names, and refused when no mode is named at all, which
the format forbids - and a test that enshrined the accepting behaviour is
now the test that pins the refusal.

Three assertions that could never fire are gone: the layer below already
refuses a number that is not finite. The material half of the fuzz target
dispatches on the container magic the way the reader does, so the
nineteen container seeds it used to walk past now reach it, and the
pairing function it never called is attacked too.

Smaller: an enum tag is compared without building a string, the table is
sized from what was parsed, two members are fetched once rather than
twice, glTF ids are stored at the width the format states them in, and
the crate root no longer makes one of two material vocabularies the
default spelling of the word.

+++ crates/mesh/src/pbr.rs
@@ -6,8 +6,14 @@
 //! vocabulary — ambient, diffuse, specular, a specular exponent — because
 //! that is what the format stores, and [`crate::mtl::Material`] holds it.
 //! A glTF material is metallic-roughness: a base colour, how metallic the
-//! surface is, how rough, and a set of maps that modulate those. The two
-//! describe surfaces in genuinely different terms.
+//! surface is, how rough, and a set of maps that modulate those.
+//!
+//! **Parts of the two do line up.** An emissive colour is an emissive
+//! colour in both, and a diffuse colour and a base colour are the same
+//! quantity under two names. It is the specular half that has no
+//! non-heuristic mapping — and a conversion carrying only the members
+//! that do correspond would quietly drop the rest, which is worse than
+//! not converting.
 //!
 //! # Why they are not merged, and never converted
 //!
@@ -44,9 +50,15 @@
 //!
 //! glTF's ranges are not a convention. The schema states `minimum` and
 //! `maximum` on the factors, so a value outside them is a document that
-//! does not conform, and the reader says so and names the member. The two
-//! readers differ because the two formats differ, not because they
-//! disagree about clamping.
+//! does not conform, and the reader says so and names the member —
+//! including the alpha cutoff, which is bounded below whatever the mode
+//! does with it. The two readers differ because the two formats differ,
+//! not because they disagree about clamping.
+//!
+//! **The check happens at the document's own width.** A number a hair
+//! past the bound narrows to exactly the bound, so a reader that
+//! converted first and checked afterwards would clamp while saying it
+//! refuses.
 
 /// Which texture a map names, and which coordinate set it reads.
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
@@ -55,9 +67,19 @@ pub struct TextureRef {
     ///
     /// **Required**, unlike almost everything else here: a map that names
     /// no texture is not a map.
-    pub texture: usize,
+    ///
+    /// A `u32` because that is what the format's own id type is, and
+    /// because five of these live in every material — widening them
+    /// to a pointer would make the size of a material depend on the
+    /// target, which is a dependency this crate keeps out of its counts.
+    ///
+    /// **Bounded, not resolved.** This crate does not read the `textures`
+    /// table, so the number is checked against that table's length and no
+    /// further: it is known to name a row that exists, and nothing here
+    /// can say what is in it.
+    pub texture: u32,
     /// Which `TEXCOORD_n` attribute to read it with. Zero by default.
-    pub uv_set: usize,
+    pub uv_set: u32,
 }
 
 /// A normal map, and how far it leans.
@@ -83,9 +105,11 @@ pub struct OcclusionTexture {
 /// How a material's alpha is meant to be read.
 ///
 /// **Three answers, and the cutoff belongs to exactly one of them.** A
-/// material that is not masked has no cutoff — the format carries the
-/// number regardless, and putting it on the variant that uses it is what
-/// stops a caller reading a threshold that means nothing.
+/// material that is not masked has no cutoff. The format lets a document
+/// state the number under any mode it declares and requires two of the
+/// three to ignore it — and forbids it outright when no mode is named
+/// at all. Putting it on the variant that uses it is what stops a caller
+/// reading a threshold that means nothing.
 #[derive(Clone, Copy, Debug, Default, PartialEq)]
 pub enum Alpha {
     /// The default: the alpha channel is ignored and the surface is
